+++ blackmamba/script/jump_to_definition.py
@@ -80,7 +80,7 @@ def _select_location(definitions):
         _open_and_scroll(definition.module_path, definition.line)
 
     v = PickerView()
-    v.name = 'Multiple definitions found'
+    v.name = '{} definitions'.format(definitions[0].name)
     v.datasource = LocationDataSource(definitions)
 
     v.shift_enter_enabled = False
@@ -89,7 +89,7 @@ def _select_location(definitions):
         '\n'
         'Esc - close • Cmd . - close with Apple smart keyboard'
     )
-    v.textfield.placeholder = 'Start typing to filter files...'
+    v.textfield.placeholder = 'Start typing to filter locations...'
     v.did_select_item_action = open_location
     v.present('sheet')
     v.wait_modal()