+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## master (unreleased)
 
 * Find usages contains symbol name in the dialog title
+* Jump to definition contains symbol name in the dialog title
 
 
 ## 1.0.2 (2017-10-01)