+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Find usages contains symbol name in the dialog title
 * Jump to definition contains symbol name in the dialog title
+* Show documentation contains symbol name in the dialog title
 
 
 ## 1.0.2 (2017-10-01)