+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Show documentation contains symbol name in the dialog title
 * `tab.open_file` has new `line` argument
     * Jump to definition, Find usages utilizes `tab.open_file` instead of `editor`'s one
+* `flake8` support added
+    * See [documentation](TODO)
 
 
 ## 1.0.2 (2017-10-01)