+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## master (unreleased)
 
+*
+
+
+## 1.1.0 (2017-10-05)
+
 * Find usages contains symbol name in the dialog title
 * Jump to definition contains symbol name in the dialog title
 * Show documentation contains symbol name in the dialog title